Move experience to component
Instead of setting the exp in the PotionType, we can set it in the
components and calculate the experience in the type based on the
components.
This means if exp rates are changed later we can more easily update it.

Also changes the experience numbers to be 10x less, not sure why they were so high